app1/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
appnope==0.1.3
asttokens==2.2.1
backcall==0.2.0
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.4
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
envyaml==1.10.211231
exceptiongroup==1.1.2
executing==1.2.0
fastapi==0.100.0
h11==0.14.0
idna==3.4
ipykernel==6.24.0
ipython==8.14.0
jedi==0.18.2
jupyter_client==8.3.0
jupyter_core==5.3.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.25.1
packaging==23.1
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.8.1
prompt-toolkit==3.0.39
psutil==5.9.5
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.0.2
pydantic_core==2.1.2
Pygments==2.15.1
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.0
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.18
stack-data==0.6.2
starlette==0.27.0
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
uvicorn==0.22.0
wcwidth==0.2.6
